==> 1.DC crime Data set up shared.ipynb <==
# Project 5: The Lab @ DC Consulting Project

The[Lab @ DC](http://thelab.dc.gov/) is an organization within [DC Mayor Muriel Bowser's](https://mayor.dc.gov/biography/muriel-bowser) office that is responsible for testing and evaluating policies and providing analysis to agencies within the DC government. In a word, The Lab is DC's own data science shop.

[Eric Foster-Moore](http://thelab.dc.gov/eric.html), a data scientist at The Lab, will help us kick off this project on Thursday, June 14th. He will speak about the types of problems that The Lab seeks to address for the city, the specific questions which are formulated and answered by data scientists at The Lab, and the ways in which their findings are communicated to stakeholders.

## Project Title: Washington DC Crime Prediction

### Authors: Eric  Plog, James Linek, Awab Idris,Kiros Gebremariam
- Cohorts of the Data Science Immersive, General Assembly @ Washington DC campus

In this notebook, we pull the raw data from Open Data DC and the Metropolitan Police Department (MPD) and transform and clean them for our analysis. **This is notebook 1 of 5.**

# Import the libraries we will need
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

import matplotlib.pyplot as plt
import math

import seaborn as sns
sns.set(font_scale=2)
cmap = sns.diverging_palette(220, 10, as_cmap=True) # one of the many color mappings

 The Data set chosen for this  DC lab project is the 2008-2017 Washington DC crime incident  report inspired from opendataset found at http://opendata.dc.gov/datasets/crime-incidents-in-2017. The dataset was obrained following the steps found in the opendataDc.gov. This site allowed us to download data by year which has been organized by PSA,political wards,ANC for all the Eight wards as csv files. These individual yearly CSVs were the concatenated together in to a single file for use. this data set contains 378,845 entries and 25 attributes that are both continuos and discrete.  The crime dataset is published by the Washington DC Metropolitant Police Department to inform the DC residents with crime trends.

# Read in the crime data from the each single year CSV file
df1 = pd.read_csv('./input/Crime_Incidents_in_2012.csv')
df2 = pd.read_csv('./input/Crime_Incidents_in_2013.csv')
df3 = pd.read_csv('./input/Crime_Incidents_in_2014.csv')
df4 = pd.read_csv('./input/Crime_Incidents_in_2015.csv')
df5 = pd.read_csv('./input/Crime_Incidents_in_2016.csv')
df6 = pd.read_csv('./input/Crime_Incidents_in_2017.csv')
df7 = pd.read_csv('./input/Crime_Incidents_in_2018.csv')

# # compiling all the CSV's in to a single pandas DF
COLUMN_NAMES = ['X', 'Y', 'CCN', 'REPORT_DAT', 'SHIFT', 'METHOD', 'OFFENSE', 'BLOCK', 'XBLOCK',
                 'YBLOCK', 'WARD', 'ANC', 'DISTRICT', 'PSA','NEIGHBORHOOD_CLUSTER', 'BLOCK_GROUP', 'CENSUS_TRACT',
                 'VOTING_PRECINCT', 'LATITUDE', 'LONGITUDE', 'BID', 'START_DATE','END_DATE', 'OBJECTID']
df = pd.DataFrame(columns=COLUMN_NAMES)
print(df.columns)   # checking consistency of column names
print(df1.columns)
print(df2.columns)
print(df3.columns)
print(df4.columns)
print(df5.columns)
print(df6.columns)
print(df7.columns)
df =pd.concat([df1,df2, df3,df4,df5,df6,df7],ignore_index =True)  # put the  DataFrames together
df

df.columns

df.head(2)

df.info()

df.to_csv('./input/dcrime.csv')
