==> 2.DC crime EDA1.ipynb <==
## DATA Cleaning and EDA

# Import the libraries we will need
import pandas as pd
import numpy as np
import scipy.stats as stats
import matplotlib.pyplot as plt
import seaborn as sns
import matplotlib.pyplot as plt
import math

# setting the visual
sns.set(font_scale=2)
cmap = sns.diverging_palette(220, 10, as_cmap=True) # one of the many color mappings
sns.set_style('whitegrid')
%config InlineBackend.figure_format = 'retina'
%matplotlib inline  

df = pd.read_csv('./input/dcrime.csv')
ward_data =pd.read_csv('./input/warddata.csv')

print (df.duplicated().sum())

# Checking Data Integrity

print(df.shape)
print(df.head(2))
print(ward_data.shape)
print(ward_data.head(2))

df.isnull().sum().sort_values(ascending=False)  

# checking the null values of the df and sorting them as percentage of the DF shape
(df.isnull().sum()/df.shape[0]).sort_values(ascending=False)  

ward_data.isnull().sum()/len(ward_data)

####  Feature Summary Statistics

###### Count of crimes committed per duty shift
The Evening shift had the highest count as compared to the three categories followed by the Day shift. However, this doesn't take into account the severity of the crime.

print('==================================================')
print( "---===<<< Duty Shift: variable = 'SHIFT' >>>===--")
var_count = df.groupby('SHIFT')
print( var_count.CCN.count())
print('================================================')

###### Count of crimes committed per Police District
The number of crimes is evenly distributed across all the PSA areas ignoring the severity of the crime.

print('=========================================================================')
print("---==< Crimes committed per police district: variable = 'DISTRICT' >==---")
var_count = df.groupby('DISTRICT')
print(var_count.CCN.count())
print('=========================================================================')

Below is a sample output from a single record to get a feel for the type of data.

print('=========================================================================')
print("---==< Crimes|OFFENSES committed per BLOCK: variable = 'BLOCK' >==---")
print(df.groupby('BLOCK').OFFENSE.value_counts())
print('=========================================================================')

print('======================================================')
print('--====< Example Record from Data Set>============')
print(df.iloc[1234])
print('======================================================')

##### Verify data quality

##### Unique Data

print('======================================================')
print (np.count_nonzero(df['BLOCK'].unique()))
df.drop('BLOCK', axis=1, inplace=True)

print (np.count_nonzero(df['BLOCK_GROUP'].unique()))
df.drop('BLOCK_GROUP', axis=1, inplace=True)
print('======================================================')

Block has 14,245 unique values indicating the street and block identifier. This value appears to be based on the CENSUS_TRACT variable, but with higher resolution. Mapping these unique values is time consuming and doesn't provide meaningful value for for further Analysis. Therefore, the column will be removed from the working data set.
Block Group has 451 unique values indicating the land management tract identifier. Mapping these unique values is time consuming and doesn't provide meaning value for further Analysis. Therefore, the column will be removed from the working data set.

# if END_DATE is NaN, then use START_DATE
df['END_DATE'].fillna(df['START_DATE'], inplace=True)

# if VOTING_PRECINCT is NaN, then set it to 0
df['VOTING_PRECINCT'].fillna(0, inplace=True)

# if NEIGHBORHOOD_CLUSTER is NaN, then set it to 0
df['NEIGHBORHOOD_CLUSTER'].fillna(0, inplace=True)

# if CENSUS_TRACT is NaN, then set it to 0
df['CENSUS_TRACT'].fillna(0, inplace=True)
# if WARD is NaN, then use ANC
# df['WARD'].fillna(df['ANC'], inplace=True)
# if WARD is NaN, then set it to 0
df['WARD'].fillna(0, inplace=True)

#  Create a dataframe that holds the central location of each Police Service Area (PSA)
#  The PSAs: according to the information from the DC metropolitan ionformation and opendata DC
# PSA's are smaller than the Police Districts, so we should have better accuracy in identifying the associated PSA
#  The PSA ID contains the District ID, so we can impute the District from the PSA
psa_loc = pd.DataFrame(df[['PSA','XBLOCK','YBLOCK']].groupby('PSA').median())

#  ---==< Estimate PSA membership based on proximity to each area's centroid >==---
def nearbyPSA(nPSA,dX,dY):
    # Default to the current PSA ID
    nearbyPSA = nPSA
    
    # Only operate on missing IDs
    if (pd.isnull(nPSA)):
        minDist = 9e99  # Set the initial closest distance to be a large value
        nearbyPSA = 0
        
        # Loop through the records in the psa_loc dataframe
        for PSA_ID, PSA in psa_loc.iterrows():
            # Calculate the distance between the report and the current PSA using the Eucleadian distance
            thisDist = math.sqrt((dX - PSA['XBLOCK'])**2 + (dY - PSA['YBLOCK'])**2)
            
            # If this distance is smaller than the current minimum distance, update the minimum distance
            if (thisDist < minDist):
                minDist = thisDist # Replace the minimum distance with the current distance
                nearbyPSA = PSA_ID # Remember which PSA this is related to
                
    # Return the ID for the closest PSA
    return [nearbyPSA, int(nearbyPSA / 100)]

#  Impute the missing PSA
df['PSA_ID'] = 0
df['DistrictID'] = 0
df[['PSA_ID','DistrictID']] = list(map(nearbyPSA,df['PSA'],df['XBLOCK'],df['YBLOCK']))

#  Display the results of this imputation method.
print('==========================================================')
print(df[['PSA','DISTRICT','PSA_ID','DistrictID']][df['PSA'].isnull()])
print('==========================================================')

##### Checking for duplicated  data

# do we have duplicate rows?
# the great thing is that i took care of them while concatenating the different values in the initial Data setup
print(df.duplicated().sum())

df.WARD.isnull().sum()

#### Feature Reeengineering 

# strip 'Precinct ' from VOTING_PRECINCT values
# http://stackoverflow.com/questions/13682044/pandas-dataframe-remove-unwanted-parts-from-strings-in-a-column
df['VOTING_PRECINCT'] = df['VOTING_PRECINCT'].apply(str).map(lambda x: x.lstrip('Precinct '))

# strip 'Cluster ' from NEIGHBORHOOD_CLUSTER values
df['NEIGHBORHOOD_CLUSTER'] = df['NEIGHBORHOOD_CLUSTER'].apply(str).map(lambda x: x.lstrip('Cluster '))

print('======================================================')
# convert REPORT_DAT to datetime
df['REPORT_DAT'] = pd.to_datetime(df['REPORT_DAT'])

# convert SHIFT to int
shift_mapping = {'day':1, 'evening':2, 'midnight':3}
df['SHIFT_Code'] = df['SHIFT'].str.lower().map(shift_mapping).astype('category')

# convert OFFENSE to numeric
# Python for Data Analysis, pg. 279
offense_mapping = {'theft/other':1, 'theft f/auto':2, 'burglary':3, 'assault w/dangerous weapon':4, 'robbery':5,
                  'motor vehicle theft':6, 'homicide':7, 'sex abuse':8, 'arson':9}
df['OFFENSE_Code'] = df['OFFENSE'].str.lower().map(offense_mapping).astype('category')
df['OFFENSE'] = df['OFFENSE'].str.replace('DANGEROUS WEAPON', 'DW')

# convert METHOD to numeric
method_mapping = {'others':1, 'gun':2, 'knife':3}
df['METHOD_Code'] = df['METHOD'].str.lower().map(method_mapping).astype('category')

# convert DISTRICT to numeric
df['DistrictID'] = df['DistrictID'].astype(np.int64)

# convert PSA to numeric
df['PSA_ID'] = df['PSA_ID'].astype(np.int64)

# convert WARD to numeric
df['WARD'] = df['WARD'].astype(np.int64)

# convert ANC to numeric
anc_mapping = {'1B':12, '1D':14, '1A':11, '1C':13, '6E':65, '4C':43, '5E':55, '2B':22, '2D':24, '2F':26, '2C':23,
       '2E':25, '2A':21, '3C':33, '3E':35, '3B':32, '3D':34, '3F':36, '3G':37, '4A':41, '4B':42, '4D':44,
       '5A':51, '5D':54, '5C':53, '5B':52, '6A':61, '6C':63, '6B':62, '6D':64, '7D':74, '7C':73, '7E':75,
       '7B':72, '7F':76, '8A':81, '8B':82, '8C':83, '8D':84, '8E':85}
df['ANC'] = df['ANC'].map(anc_mapping).astype('category')

# convert NEIGHBORHOOD_CLUSTER to numeric
df['NEIGHBORHOOD_CLUSTER'] = df['NEIGHBORHOOD_CLUSTER'].astype(np.int64)

# convert CENSUS_TRACT to numeric
df['CENSUS_TRACT'] = df['CENSUS_TRACT'].astype(np.int64)

# convert VOTING_PRECINCT to numeric
df['VOTING_PRECINCT'] = df['VOTING_PRECINCT'].astype(np.int64)

# convert CCN to numeric
df['CCN'] = df['CCN'].astype(np.int64)

# convert XBLOCK, YBLOCK to numeric
df['XBLOCK'] = df['XBLOCK'].astype(np.float64)
df['YBLOCK'] = df['YBLOCK'].astype(np.float64)

# convert START_DATE, END_DATE to dateime
df['START_DATE'] = pd.to_datetime(df['START_DATE'])
df['END_DATE'] = pd.to_datetime(df['END_DATE'])
print('======================================================')
print(df.info())
print()
print(df.iloc[1234])
print('======================================================')


##### How do we identify Outliers
one hundredsix crimes were reported in 2012 which started before 2011. Six of the crimes started in 1915. Are these start dates in error or were these unsolved crimes recently solved through new technology advancements? Is it truly a 100-year-old cold case, or did the person simply enter the wrong century? beacuse the timestamp shows from 01/16/1800 one crime, two crimes from 1912 reported in 2011 and one crime of 1914 reported in 2011.

print('======================================================')
print('---======< Dealing with outliers>=======-------')
print(df['START_DATE'][df['START_DATE']<'1/1/2011'].count())
print(sorted(df['START_DATE'][df['START_DATE']<'1/1/2011'])[:10])
print('======================================================')

four hundred fourty Five crimes were reported in 2013 which started before 2012. Six of the crimes started in 1915. Are these start dates in error or were these unsolved crimes recently solved through new technology advancements? Is it truly a 100-year-old cold case, or did the person simply enter the wrong century? beacuse the timestamp shows from 01/16/1800 one crime, two crimes from 1912 reported in 2011 and one crime of 1914 reported in 2012.

print('======================================================')
print('---======< Dealing with outliers>=======-------')
print(df['START_DATE'][df['START_DATE']<'1/1/2012'].count())
print(sorted(df['START_DATE'][df['START_DATE']<'1/1/2012'])[:10])
print('======================================================')

35633 crimes were reported in 2014 which started before 2013. Six of the crimes started in 1915. Are these start dates in error or were these unsolved crimes recently solved through new technology advancements? Is it truly a 100-year-old cold case, or did the person simply enter the wrong century? beacuse the timestamp shows from 01/16/1800 one crime, two crimes from 1912 reported in 2011 and one crime of 1914 reported in 2013.

print('======================================================')
print('---======< Dealing with outliers>=======-------')
print(df['START_DATE'][df['START_DATE']<'1/1/2013'].count())
print(sorted(df['START_DATE'][df['START_DATE']<'1/1/2013'])[:10])
print('======================================================')

###### Creating New Features
creating new feature CRIME_TYPE to indicate if the crime was violent or nonviolent. This feature will be used for later prediction purposes.

# creating feature for crime type 1 = Violent, 2 = non violent
violent_offense = [4, 5, 7, 8]
df['CRIME_TYPE'] = np.where(df['OFFENSE_Code'].isin(violent_offense), 1, 2)
df['CRIME_TYPE'] = df['CRIME_TYPE'].astype('category')

Creating additional feature AGE to indicate the timespan between the latest the crime could have been committed and earliest.

# create age of crime END_DATE - START_DATE in seconds
df['AGE'] = (df['END_DATE'] - df['START_DATE'])/np.timedelta64(1, 's')

# create new feature TIME_TO_REPORT to indicate the timespan between the latest time the crime was committed and
# the time it was reported to the MPD. that is  time it took from crime to report it, REPORT_DAT - END_DATE in seconds
df['TIME_TO_REPORT'] = (df['REPORT_DAT'] - df['END_DATE'])/np.timedelta64(1, 's')

df['DATE'] = pd.to_datetime(df['END_DATE'], format = '%d/%m/%Y %H:%M:%S')

def date_separate(df):
    """ function that separates the sepecific interest days,weeks from the END_DATE"""
    df = df.copy()
    df['Year'] = pd.DatetimeIndex(df['DATE']).year
    df['Month'] = pd.DatetimeIndex(df['DATE']).month
    df['Day'] = pd.DatetimeIndex(df['DATE']).day
    df['hour'] = pd.DatetimeIndex(df['DATE']).hour
    df['dayofyear'] =pd.DatetimeIndex(df['DATE']).dayofyear
    df['week'] =pd.DatetimeIndex(df['DATE']).week
    df['weekofyear'] =pd.DatetimeIndex(df['DATE']).weekofyear
    df['dayofweek'] =pd.DatetimeIndex(df['DATE']).dayofweek
    df['weekday'] =pd.DatetimeIndex(df['DATE']).weekday
    df['quarter'] =pd.DatetimeIndex(df['DATE']).quarter
    return df

# here i have created an additional columns that helps us in analyzing the timing of the crimes happeing either by day
# week, month, year as well as the quarters and weekdays and week ends.
df['Year'] = pd.DatetimeIndex(df['DATE']).year
df['Month'] = pd.DatetimeIndex(df['DATE']).month
df['Day'] = pd.DatetimeIndex(df['DATE']).day
df['hour'] = pd.DatetimeIndex(df['DATE']).hour
df['dayofyear'] =pd.DatetimeIndex(df['DATE']).dayofyear
df['week'] =pd.DatetimeIndex(df['DATE']).week
df['weekofyear'] =pd.DatetimeIndex(df['DATE']).weekofyear
df['dayofweek'] =pd.DatetimeIndex(df['DATE']).dayofweek
df['weekday'] =pd.DatetimeIndex(df['DATE']).weekday
df['quarter'] =pd.DatetimeIndex(df['DATE']).quarter

#### MOST IMPORTANT ATTRIBUTTES

###### OFFENSE

print('==============================================================')
print('---<Examine the frequency of types of crimes/offenses>--======')
print('    __________________________________________________')
print(' Total Offenses - Count')
total_crime = df.CCN.count()
print (total_crime)
print('')
crime_rate = df.groupby('OFFENSE')
print('_________________________________')
print (' Offense Type - Count')
print('_________________________________')
print (crime_rate.CCN.count())
print ('')
print('_________________________________')
print (' Offense as percentage of total ')
print ('_______________________________')
print (crime_rate.CCN.count() / total_crime * 100.0)
print()
print ('___________________________')
print (' Offense Rate per 100,000 ')
print ('__________________________')
print (crime_rate.CCN.count() / 6.93972 )
print()
print ('_____________________________________')
print (' Odds of being a victim - by offense')
print ('____________________________________')
print (crime_rate.CCN.count() / 693972)

print('==============================================================')

###### METHOD

print('==============================================================')
print('-<Examine the frequency of types of crimes/offenses by Method>===')
print ('_____________________________________')
print(' Total Methods - Count') 
print ('_____________________________________')
print(df.CCN.count())
print ('')

method_rate = df.groupby('METHOD')
print ('_____________________')
print ('Method Type - Count')
print ('____________________')
print(method_rate.CCN.count())
print()
print ('______________________________')
print (' Method as percentage of total ')
print ('_____________________________')
print (method_rate.CCN.count() / total_crime * 100.0)

print ('_________________________')
print (' Method Rate per 100,000 ')
print ('_________________________')
print( method_rate.CCN.count() / 6.93972)
print ('___________________________________')
print (' Odds of being a victim - by method')
print ('_______________________________')
print (method_rate.CCN.count() / 693972)
print('==============================================================')

##### Age of Crime


print('==============================================================')
print('-<Examine the age of  crimes/offenses by Time >===')

# hours
temp = df['AGE'] / 3600

print( temp.describe())
print('') 

# exclude temp two std away from mean
#http://stackoverflow.com/questions/23199796/detect-and-exclude-outliers-in-pandas-dataframe
print (temp[~(np.abs(temp - temp.mean())>(2*temp.std()))].describe())
print('==============================================================')


### Time to Report a Crime

print('==============================================================')
print('-<Examine the frequency of time it takes to report a of crimes/offenses >===')
# hours
temp = df['TIME_TO_REPORT'] / 3600

print(temp.describe())
print ('-----------------------')

# excluding temp two std away from mean
print (temp[~(np.abs(temp - temp.mean())>(2*temp.std()))].describe())
print('==============================================================')

df.to_csv('./input/dc_crime2.csv')
