==> 4.Modeling and Evaluation DC Crime.ipynb <==
## MODELING AND EVALUATION


## Load Necessary Library and Data

# import Necessary Libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from __future__ import print_function
from pandas import read_csv
from pandas.tools.plotting import parallel_coordinates
from sklearn import metrics
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from sklearn.decomposition import PCA
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import VotingClassifier
from sklearn.linear_model import SGDClassifier
# Visual setting
sns.set(font_scale=1)
cmap = sns.diverging_palette(220, 10, as_cmap=True) # one of the many color mappings
plt.style.use('fivethirtyeight')
sns.set_style('whitegrid')
plt.style.use('ggplot')
%config InlineBackend.figure_format = 'retina'
%matplotlib inline

df=pd.read_csv('./input/dc_crime3.csv')
ward_data =pd.read_csv('./input/warddata.csv')

df.head()[0:2]

## X  Features and y target variable

# Features =['SHIFT', 'OFFENSE', 'METHOD', 'DISTRICT', 'PSA', 'WARD', 'CENSUS_TRACT', 
#                'VOTING_PRECINCT', 'CRIME_TYPE', 'CCN', 'XBLOCK', 'YBLOCK', 'START_DATE', 'DistrictID', 'SHIFT_Code', 
#                'METHOD_Code', 'AGE', 'TIME_TO_REPORT']

# # please choose features of your interest such as Age, time, crime type, unemployment, Housing price
# X = df.drop('Features')
# y= df[Features]

# # create ANC attribute names and values
# y = df['OFFENSE_Code']
# X = df.drop('OFFENSE_Code',axis =1)

# from sklearn.preprocessing import Imputer

# X_train,X_test,y_train,y_test = train_test_split(X,y,stratify =y,random_state=1837)
# imp= Imputer()
# X_train = imp.fit_transform(X_train)
# X_test = imp.transform(X_test)

## Creating & Training KNN Model

# from sklearn.neighbors import KNeighborsClassifier
# knn = KNeighborsClassifier(n_neighbors=10)
# knn.fit(X_train,y_train)

# knn.score(X_test,y_test)

# knn.score(X_train,y_train)

## Creating & Training Decision Tree Model

# from sklearn.tree import DecisionTreeClassifier
# dtree = DecisionTreeClassifier(max_depth=50, random_state=42)

# dtree.fit(X_train,y_train)

# y_pred=dtree.predict(X_test)

# dtree.score(X_test,y_test)

# dtree.score(X_train,y_train)

# y_pred

# treefeatures=dtree.feature_importances_
# indices = np.argsort(treefeatures)

# treefeatures

## Tree Visualization

# feature_names=[ 'dayofweek', 'dayofyear', 'hour', 'month', 'week','latitude', 'longitude']

## Creating & Training Random Tree Model

# from sklearn.ensemble import RandomForestClassifier
# rfc = RandomForestClassifier(n_estimators=100)
# rfc.fit(X_train, y_train)

# y_pred=rfc.predict(X_test)

# rfc.score(X_test,y_test)

# rfc.score(X_train,y_train)

# Think of doing  Confusion Matrix
